pyproject.toml
[tool.poetry]
name = "seared-quail"
version = "0.1.0"
description = "A digital restaurant menu web application."
authors = ["Lucas Connors <lucas@revolutiontech.ca>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.8"
Django = "~2.1.10"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.5"
gevent = "^20.5.0"
psycogreen = "^1.0.2"
Pillow = "^7.1.2"
django-ordered-model = "^3.3.0"
gevent-socketio = {git = "https://github.com/abourget/gevent-socketio.git", rev = "1cdb1594a315326987a17ce0924ea448a82fab01"}
python-dotenv = "^0.13.0"
django-configurations = "^2.2"

[tool.poetry.dev-dependencies]
coverage = "^5.1"
django-pigeon = "^0.3.0"
black = "^19.10b0"
pyupgrade = "^2.3"
isort = "^4.3.21"

[tool.isort]
# https://github.com/psf/black/blob/master/docs/compatible_configs.md#isort
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.black]
line-length = 88
target_version = ["py38"]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
